Fixed a bunch of errors and tried a different font.

+++ Gene/Engine/Private/Platform/Win32/Win32Time.cc
@@ -7,6 +7,6 @@
 
     void GameTime::Sleep(float milliseconds)
     {
-        ::Sleep(milliseconds);
+        ::Sleep(static_cast<DWORD>(milliseconds));
     }
 #endif
